test: enforce boundary-only testing, remove redundant tests
Consolidate module tests to one file per boundary (drinks.test.ts,
identity.test.ts). Expose purgeSearchCache and navigation helpers
through public seams so tests never import module internals. Remove
admin route action tests that duplicated service + routeAction coverage.
Keep route tests that guard cache header policy.
